+++ src/main/ruby/truffleruby/core/immediate.rb
@@ -51,10 +51,6 @@ module ImmediateValue
     []
   end
   private :protected_singleton_methods
-
-  def clone
-    raise TypeError, "can't clone #{self.class.name}"
-  end
 end
 
 class NilClass